convert raster tile cache asset into submodule
