+++ app/models/pydantic/creation_options.py
@@ -95,6 +95,8 @@ class RasterTileSetAssetCreationOptions(BaseModel):
     subset: Optional[str]
     grid: Grid
     symbology: Optional[Symbology] = None
+    compute_stats: bool = True
+    compute_histogram: bool = True
 
     class Config:
         extra = "forbid"
