+++ app/tasks/raster_tile_cache_assets/__init__.py
@@ -1,0 +1,4 @@
+from .raster_tile_cache_assets import (
+    raster_tile_cache_asset,
+    raster_tile_cache_validator,
+)
